Refactor: Split utils.py into feature-specific modules
I've decomposed `api/web_hook/utils.py` into two new modules:
- `api/web_hook/wiki_data_processor.py`: Contains functions related to wiki data processing and formatting (`extract_wiki_structure_xml`, `parse_wiki_structure`, `clean_and_format_content`, `generate_llms_txt`).
- `api/web_hook/wiki_exporter.py`: Contains the `export_wiki_python` function for exporting wiki content.

Each new module includes an `if __name__ == '__main__':` block with tests relevant to its functions, migrated from the original `utils.py`.
Dependencies `aiohttp` and `pydantic` were identified and are assumed to be part of your project's requirements now.

The `api/web_hook/wiki_generation_service.py` file has been updated to import necessary functions from `wiki_data_processor.py`.

The original `api/web_hook/utils.py` file has been removed.

This refactoring improves modularity and clarity by separating distinct features into their own files, making the codebase easier to understand and maintain, while preserving all original functionality and testing capabilities.
